Logging Pane Re-work First Pass
RCSpeedo2 is going to get rid of the notion of "logs" altogether.
Instead, "models" will be created by the user. He/she will then fly
a set of passes that are recorded and pruned as necessary. Once
the user is satistfied, the set of passes will be saved to the model.

The log manager will now allow the user to adjust the models that
he/she can save to. Logs can be exported by model. Eventually,
I would like to add the ability to view logs in-app.

This patch removes the old notion of "log names".